PEL: Change info trace to a debug one
The PLDM daemon was missing the call to HostReject when PHYP responded
to getting a new PEL with the 'file discarded' response.  Now that it's
being added, this trace was showing up multiple times as PHYP would
accept one new PEL at a time as the OS slowly acked them.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I13bccf821b96a8d918d33646893efa34faf4c088