PEL: Change asprintf call to snprintf
A static analysis security scanner (HCL AppScan SAST) flagged the
asprintf call as a vulnerability because the format string is passed in.

While really it isn't an issue because the format string is always
hardcoded in the calling function and isn't input by a user, this commit
changes it anyway to get it off the list.

Use snprintf instead.  While it still takes a passed in format string, a
hardcoded maximum length is used so that the scanner shouldn't flag it
as a possible buffer overrun.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: Ife309b63470536940ac88c27d13fe73716096326