PEL: Update sev of critical threshold errors
The Temperature/Voltage CriticalHigh/Low PELs previously had a PEL
severity of 'predictive', and this commit changes them to
'unrecoverable' so that it matches the D-Bus severity value it's being
created with.

Note: The D-Bus severity value is 'critical'.  While there is a PEL
severity value of 'critical', both 'unrecoverable' and 'critical' get
mapped to the Redfish field value of 'Critical' anyway, and the PEL
'critical' field seems to be just used for system crashes and the like.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I174c48180abe9ddca09d0e020e6d16f81311bb61