PEL: Add the DRAM manufacturer info of DIMM callout in UD section
Add the called out DIMMs DRAM manufacturer info to the user data
section of the PEL to assist the service engineers in identifying the
manufacturer of the faulty DRAMs packaged within the DIMM module
directly from the logs, aiding in quick resolution.

The changes also moves the pdbg target and libekb initialization to
the PEL startup which avoids the need of multiple initialization as
the existing design.

When a PEL calls out a DIMM FRU, the DRAM manufacturer ID and the
expanded location code of those DIMMs are added to the SysInfo user
data section  of the generated PEL in JSON format under the key 'DIMMs
Additional Info'.

In case of  any errors occur during the collection or processing of
the manufacturer data, the error messages will be logged in the 'PEL
Internal Debug Data' section as a JSON array under the key 'DIMMs Info
Fetch Error' as a separate user data section.

Tested :

Below is a portion of PEL(callout section and User Data section are
shown) which callout the DIMM P0-C32.

```
"Hex Word 9":               "00000000",
    "Callout Section": {
        "Callout Count":        "1",
        "Callouts": [{
            "FRU Type":         "Normal Hardware FRU",
            "Priority":         "Mandatory, replace all with this
                                 type as a unit",
            "Location Code":    "UXXX.YYY.WWW004A-P0-C32",
            "Part Number":      "7777777",
            "CCIN":             "1234",
            "Serial Number":    "YYYYYY"
        }]
    }
```
"User Data": {
    "Section Version": "1",
    "Sub-section type": "1",
    "Created by": "bmc error logging",
    "BMCLoad": "0.65 0.69 0.64",
    "BMCState": "Ready",
    "BMCUptime": "0y 0d 0h 17m 43s",
    "BootState": "Unspecified",
    "ChassisState": "Off",
    "DIMMs Additional Info": [
        {
            "DRAM Manufacturer ID": [
                "0x88",
                "0xAA"
            ]
            "Location Code": "UXXX.YYY.WWW004A-P0-C32",
        }
    ],
    "FW Version ID": "fw1060.20-4-1060.2432.20240729a (NL1060_068)",
    "HostState": "Off",
    "System IM": "50001001"
}
```

Change-Id: I2ff81c66e63b99e8e84378ec78f586fb9b6322d7
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>