PEL: DRAM: Fixing the missed comma in lg2 trace
Root Cause: In commit afba316, two traces in the lg2 logging missed a
comma, leading to a compilation error.

Solution: The lg2 traces have been corrected by adding the missing
comma. Additionally, the systemd signal subscription and
unsubscription logic has been moved outside of the PEL_ENABLE_PHAL
flag, as it is not dependent on PHAL.

Change-Id: Ib8b17af95be356c6649c7804e4e46704498f3f3d
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>