lg2: use DEBUG_INVOCATION to reduce debug journal
The systemd-v257 release introduced a new convention that can be used
by daemons to reduce the amount of calls to journald for debug-level
messages: DEBUG_INVOCATION[1].  Daemons will typically start with
DEBUG_INVOCATION not set, but in the event of a crash, the daemon will
be restarted with it set.  Daemons are expected to emit higher debug
messages when ran in that mode.

Adjusting the logger code to avoid sending debug messages to journald
unless DEBUG_INVOCATION is set.  This reduces the overhead of logging
by skipping the journald calls entirely.

[1]: https://github.com/systemd/systemd/commit/7d8bbfbe0852ec89590d1dc5e28afc95d6d44fa1

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I5896c4714b92786c8f859f2e5ab389753e4b5c26