PELs: Use pldm_msg_hdr_correlate_response()
This call to ensure the PLDM response matches the PLDM request for the
'New PEL Available' command was previously missing (since it didn't
exist at the time this code was written), causing the code to think it
got the correct response when really it was someone else's.

It would then move on to the next PEL before it actually heard back
about the current one, causing issues on the host side.

Change-Id: I471d7727a2b8c77a0ffc85c15cef7531898d22d7
Signed-off-by: Matt Spinler <spinler@us.ibm.com>