PEL: Ensure PLDM response valid before using it (#90)
Make sure that pldm_transport_recv_msg() was successful before passing
the response value it returns into pldm_msg_hdr_correlate_response().
Otherwise, a null pointer may be passed into it which will cause a
crash.

Also, instead of manually freeing the response on the function exit
points, just wrap it in a small Response struct that will free it in its
destructor.

Change-Id: I213c44758c248d14d83144a7bad397e84fe5602b

Signed-off-by: Matt Spinler <spinler@us.ibm.com>