entry: remove AdditionalData2
The 'AdditionalData2' property was a temporary addition to ease the
transition from an array to a dictionary for the additional data.
All users have been migrated to use the dictionary style and from the
'AdditionalData' property, so the '2' property can be removed.

Tested: Loaded an existing log database and observed dbus objects
are created successfully and with the 'AdditionalData2' field missing.
Created a new event with `log-create` CLI and observed new entry is
created successfully in dbus.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Ica3a9e5730eca8a7cf7755b769189347220e1639