PEL: Check for Py_None from parser calls (#108)
Explicitly check if the UserData or SRC python plugin parsers return a
python None object and just return immediately.  In the UserData case
this results in a hex dump, and in the SRC case it just won't print the
SRC description.

Otherwise, peltool will crash when it tries to extract a string from the
result.

Tested:
peltool no longer crashes.

Change-Id: I2cb87a6071ecc2cbeb9df055f13039ad31fc6384

Signed-off-by: Matt Spinler <spinler@us.ibm.com>