Audit Log: Switch to use auparse library function
The Linux auparse library provides a function which handles stripping
quotes from the value part of an audit field. [1] This function also
handles interpreting user or group ID numbers to names. Switch to use
this function instead of the local getValue() function.

Tested: Confirmed a variety of parsed audit entries before and after the
change have stripped the quotes in the same manner.

[1] https://linux.die.net/man/3/auparse_interpret_field

Signed-off-by: Janet Adkins <janeta@us.ibm.com>